printing csv, console and txt
All three options for printing are aligned properly with thoughts on readability.
made an extra package formatters that the data which need to be printed out will be printed out in the way I want. Which means I had to implement another interface called reportformatter. It took me quite a while to understand that I needed to have another package to run through as all three data printouts would need a different way to be aligned.